Move the BlocBuilder from JobList widget to JobListPage to get rid of possible dependencies problem.
