+++ lib/widgets/job_list.dart
@@ -8,57 +8,49 @@ import 'category_container.dart';
 import 'job_card.dart';
 
 class JobList extends StatelessWidget {
-  const JobList({Key? key}) : super(key: key);
+
+  final List<Object> data;
+
+  JobList(this.data);
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<HomeBloc, HomeStates>(
-      builder: (BuildContext context, HomeStates state) {
-        if (state is OnLoading) {
-          return LinearProgressIndicator();
-        } else if (state is LoadDataFail) {
-          return Center(child: Text(state.error));
-        } else if (state is LoadDataSuccess) {
-          var data = (state).data['job_search']?['result'];
-          return SingleChildScrollView(
-            physics: ScrollPhysics(),
-            child: Column(
-              mainAxisSize: MainAxisSize.max,
-              children: [
-                Container(
-                  height: 230,
-                  width: double.infinity,
-                  child: GridView(
-                    scrollDirection: Axis.horizontal,
-                    shrinkWrap: true,
-                    padding: EdgeInsets.all(25),
-                    children: jobCategories
-                        .map((catData) => CategoryButton(catData))
-                        .toList(),
-                    gridDelegate: SliverGridDelegateWithMaxCrossAxisExtent(
-                      maxCrossAxisExtent: 150,
-                      childAspectRatio: 2 / 3,
-                      crossAxisSpacing: 20,
-                      mainAxisSpacing: 20,
-                    ),
-                  ),
-                ),
-                ListView.builder(
-                  itemCount: data.length,
-                  shrinkWrap: true,
-                  physics: NeverScrollableScrollPhysics(),
-                  itemBuilder: (context, index) {
-                    final jobId = (data[index] as Map);
-                    return JobCard(job: jobId);
-                  },
-                ),
-              ],
+    return SingleChildScrollView(
+      physics: ScrollPhysics(),
+      child: Column(
+        mainAxisSize: MainAxisSize.max,
+        children: [
+          Container(
+            height: 230,
+            width: double.infinity,
+            child: GridView(
+              scrollDirection: Axis.horizontal,
+              shrinkWrap: true,
+              padding: EdgeInsets.all(25),
+              children: jobCategories
+                  .map((catData) => CategoryButton(catData))
+                  .toList(),
+              gridDelegate: SliverGridDelegateWithMaxCrossAxisExtent(
+                maxCrossAxisExtent: 150,
+                childAspectRatio: 2 / 3,
+                crossAxisSpacing: 20,
+                mainAxisSpacing: 20,
+              ),
             ),
-          );
-        } else {
-          return Center(child: CircularProgressIndicator());
-        }
-      },
+          ),
+          ListView.builder(
+            itemCount: data.length,
+            shrinkWrap: true,
+            physics: NeverScrollableScrollPhysics(),
+            itemBuilder: (context, index) {
+              final jobId = (data[index] as Map);
+              return JobCard(job: jobId);
+            },
+          ),
+        ],
+      ),
     );
+
+
   }
 }
